Fixed autonomy logging with Unit and Integration tests.
